Modified the dynamic programming part
	1. The task length is all integers of timestamp.
	   The timestamp is assumed to be 0.1.
	   For example: if a task length is 7 at 1.0v, then at 0.8v, its
	   length should be 7/0.8 = 8.75. Then it will be round to 8.8
	   which is 88 timestamp. We always use ceil() in rounding.
	2. We invalidate all dynamic programming table entries for a timestamp
	   that is not possible compelte a set of tasks.
	   For example: we have 2 tasks and 2 speeds:
	   The length of T1 at the 2 speeds are (3, 5) and
	   The length of T2 at the 2 speeds are (4, 7).
	   So the only valid timestamp in the table entries are (7, 9, 10, 12);
	   Other entries are all nonvalid.

+++ DP/taskScheduling.hpp
@@ -9,12 +9,12 @@ class dynProg {
 	private:
 		int m_numOfTask; //Number of task
 		int m_numOfVolt;
-		double m_deadline;// The deadline of all the task
+		int m_deadline;// The deadline of all the task
 		vector<double>m_inputDuration; // The task duration on 1.0V
 		vector<double>m_inputEnergy; // The task energy on 1.0V
 
 		//double m_taskDuration[m_numOfTask][5];
-		vector< vector<double> > m_taskDuration;// The task duration from 0.8V to 1.2V
+		vector< vector<int> > m_taskDuration;// The task duration from 0.8V to 1.2V
 		//double m_taskEnergy[m_numOfTask][5];
 		vector< vector<double> > m_taskEnergy;// The task energy from 0.8V to 1.2V
 		//double m_scheduleEnergy[m_numOfTask][m_deadline * 10 + 1];
